pyproject.toml
[project]
name = "ai-employee"
version = "0.1.0"
description = "Turbo Launch AI Employee - Autonomous FTE for business automation"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth>=2.0",
    "google-auth-oauthlib>=1.0",
    "google-auth-httplib2>=0.1",
    "google-api-python-client>=2.0",
    "watchdog>=4.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    "requests>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "black>=24.0",
    "ruff>=0.1",
]
whatsapp = [
    "playwright>=1.40",
]
social = [
    "requests-oauthlib>=2.0",
]

[project.scripts]
ai-employee = "scripts.main:main"
ai-watch = "scripts.watchers.filesystem_watcher:main"
ai-orchestrate = "scripts.orchestrator.orchestrator:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]

[tool.uv]
package = true

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"
